pythonGC/pyproject.toml
[tool.poetry]
name = "pythonGC"
version = "0.1.0"
description = ""
authors = ["tomast1337 <vycasnicolas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.3"
PyOpenGL = "^3.1.6"
PySDL2 = "^0.9.14"
pysdl2-dll = "^2.24.1"
rich = "^12.6.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
